pyproject.toml
[project]
name = "web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.12.12",
    "uuid>=1.30",
]
